i am in fking loop

+++ components/posts/action.ts
@@ -3,13 +3,13 @@
 import { db } from "@/db/drizzle"
 import { post } from "@/db/schema"
 import { getServerSession } from "@/lib/getServerSession"
-import { eq } from "drizzle-orm"
+import { and, eq } from "drizzle-orm"
 
 export const deletePostAction = async (id: string) => {
     const session = await getServerSession()
     if (!session) throw new Error("Unauthorized")
 
-    const res = await db.delete(post).where(eq(post.id, id)).returning()
+    const res = await db.delete(post).where(and(eq(post.id, id), eq(post.userId, session.user.id))).returning()
     if (!res) {
         throw new Error("Post not found")
     }
